app/recipe/serializers.py
"""
Serializers for Recipe APIs
"""

from typing import Any, Dict, List

from core.models import Recipe, Tag
from rest_framework import serializers


class TagSerializer(serializers.ModelSerializer[Tag]):  # type: ignore[misc]
    """Serializer for Tag objects"""

    class Meta:
        model = Tag
        fields = ["id", "name"]
        read_only_fields = ["id"]


class RecipeSerializer(serializers.ModelSerializer[Recipe]):  # type: ignore[misc]
    """Serializer for Recipe objects"""

    tags = TagSerializer(many=True, required=False)

    class Meta:
        model = Recipe
        fields = ["id", "title", "time_minutes", "price", "link", "tags", "description"]
        read_only_fields = ["id"]

    def create(self, validated_data: Dict[str, Any]) -> Recipe:
        """create a recipe"""
        tags: List[Dict[str, Any]] = validated_data.pop("tags", [])
        recipe = Recipe.objects.create(**validated_data)
        auth_user = self.context["request"].user
        for tag in tags:
            tag_obj, created = Tag.objects.get_or_create(user=auth_user, **tag)
            recipe.tags.add(tag_obj)

        return recipe


class RecipeDetailSerializer(RecipeSerializer):
    """Serializer for Recipe detail view"""

    class Meta(RecipeSerializer.Meta):
        fields = RecipeSerializer.Meta.fields + ["description"]
