requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.15,<3.16
psycopg2>=2.9,<3.0
drf-spectacular>=0.27.0,<0.28
flake8>=7.0
black>=24.0
isort>=5.13
pre-commit>=2.20
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
coverage>=7.0,<8.0
mypy>=1.8,<2.0
ruff>=0.3,<1.0
types-requests
django-stubs
